misc/HtmlTestSuite/results/test-0556.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><title>Sticky positioned element should be observable by offsetTop and offsetLeft</title>
<link href="https://www.w3.org/TR/css-position-3/#sticky-pos" rel="help"/>
<meta content="This test checks that a sticky positioned element&#xA;should be observable by offsetTop/offsetLeft." name="assert"/>

<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>

<style>
body {
  margin: 0;
}

.container {
  position: relative; /* Required for offsetTop/offsetLeft tests. */
  overflow: scroll;
  width: 200px;
  height: 200px;
}

.spacer {
  width: 2000px;
  height: 2000px;
}

.box {
  width: 100px;
  height: 100px;
  background-color: green;
}

.sticky {
  position: sticky;
  top: 50px;
  left: 20px;
}
</style>

</head><body><div class="container" id="scroller1">
  <div class="spacer"/>
</div>

<script>
test(() =&gt; {
    var scroller = document.getElementById('scroller1');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;

    var sticky = document.createElement('div');
    sticky.className = 'sticky box';
    scroller.insertBefore(sticky, scroller.querySelector('.spacer'));

    assert_equals(sticky.offsetTop, scroller.scrollTop + 50);
    assert_equals(sticky.offsetLeft, scroller.scrollLeft + 20);
}, 'offsetTop/offsetLeft should be correct for sticky after script insertion');
</script>

<div class="container" id="scroller2">
  <div class="sticky box" id="sticky2"/>
  <div class="spacer"/>
</div>

<script>
test(function() {
    var scroller = document.getElementById('scroller2');
    var sticky = document.getElementById('sticky2');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;

    var div = document.createElement('div');
    div.style.height = '65px';
    scroller.insertBefore(div, sticky);

    assert_equals(sticky.offsetTop, scroller.scrollTop + 50);
    assert_equals(sticky.offsetLeft, scroller.scrollLeft + 20);
}, 'offsetTop/offsetLeft should be correct for sticky after script-caused layout');
</script>
</body></html>
